refactor(studio-services): split service console sections
